Add files via upload
This is the correct version. You can access my fifth page by clicking on any "*" in the cookie recipe. Sorry for the late submission. My code broke when I tried to add a href link to my "Explanations" to return to the home page, and I realized afterward that it was because I didn't correctly close a paragraph.